Fixed the functionality
Had to use some encapsulation because I forogt all about it and it helps solve so many of my problems. Have seperate methods now instead of having code stuck inside the actionListener. It seems like the size of text files isnt an issue because it scans through instantly and can now return each line where the search string is found and count the number of lines. Next step is to come up with some sort of ranking system, allow searching of multiple files at the same time and add a panel to the GUI for output.